added algorithm

+++ cpp00/ex00/megaphone.cpp
@@ -6,11 +6,12 @@
 /*   By: juramos <juramos@student.42madrid.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/04 12:55:56 by juramos           #+#    #+#             */
-/*   Updated: 2024/06/05 12:08:54 by juramos          ###   ########.fr       */
+/*   Updated: 2024/07/04 11:32:48 by juramos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <iostream>
+#include <algorithm>
 
 int	main (int argc, char **argv)
 {
@@ -21,7 +22,7 @@ int	main (int argc, char **argv)
 	if (argc == 1)
 		std::cout << "* LOUD AND UNBEARABLE NOISE *";
 	else
-		while (++i < argc)
+		while (++i < (size_t)argc)
 		{
 			s = argv[i];
 			std::transform(s.begin(), s.end(), s.begin(), ::toupper);
